created abstract functions for addDataLayer and filterLayers
