+++ src/app/@ansyn/plugins/openlayers/layers/openlayers-osm-layers.plugin.ts
@@ -4,6 +4,9 @@ import { ImageryPlugin } from '@ansyn/imagery/model/base-imagery-plugin';
 import { IMAGERY_MAP_COMPONENTS, ImageryMapComponentConstructor } from '@ansyn/imagery/model/imagery-map-component';
 import { Inject } from '@angular/core';
 import { BaseOpenlayersLayersPlugin } from '@ansyn/plugins/openlayers/layers/base-openlayers-layers-plugin';
+import { ILayer, layerPluginType } from '@ansyn/menu-items/layers-manager/models/layers.model';
+import OSM from 'ol/source/osm';
+import TileLayer from 'ol/layer/tile';
 
 
 @ImageryPlugin({
@@ -15,6 +18,28 @@ export class OpenlayersOsmLayersPlugin extends BaseOpenlayersLayersPlugin {
 	constructor(protected store$: Store<any>,
 				@Inject(IMAGERY_MAP_COMPONENTS) protected imageryMapComponents: ImageryMapComponentConstructor[]) {
 		super(store$, imageryMapComponents);
+	}
+
+	addDataLayer(layer: ILayer, groupName: string) {
+		const vectorLayer = new TileLayer({
+			zIndex: 1,
+			source: new OSM({
+				attributions: [
+					layer.name
+				],
+				opaque: false,
+				url: layer.url,
+				crossOrigin: null
+			})
+		});
+		vectorLayer.set('id', layer.id);
+		this.addGroupLayer(vectorLayer, groupName);
+	}
+
+	filterLayers(layers, selectedLayersIds): [ILayer[], string[]] {
+		const osmLayers = layers.filter(layer => layer.layerPluginType === layerPluginType.OSM);
+		const validSelected = selectedLayersIds.filter(id => osmLayers.some((layer) => layer.id === id));
+		return [osmLayers, validSelected];
 	}
 
 	// subscribers = [];
